Completed DAO

+++ WiselyWearDemo/main/java/entity/CommonWeatherFactory.java
@@ -3,7 +3,7 @@ package entity;
 public class CommonWeatherFactory implements WeatherFactory {
 
     @Override
-    public Weather create(int currentTemperature, String location) {
+    public Weather create(double currentTemperature, String location) {
         return new CommonWeather(currentTemperature, location);
     }
